Bppflag from tim2 shares values with field depth. so made the enum global so we can use it in both places.
